+++ src/gameplay/mapgen/builders.cpp
@@ -62,8 +62,8 @@ std::uint32_t position_hash(std::uint32_t seed, int x, int y,
 // A grass field of (tw x th) tiles; PixieData owns the heap buffer.
 PixieData make_grid(int tw, int th, unsigned char fill)
 {
-    auto* buf = new unsigned char[static_cast<std::size_t>(tw) * th];
-    std::fill(buf, buf + static_cast<std::size_t>(tw) * th, fill);
+    auto* buf = new unsigned char[static_cast<std::size_t>(tw) * static_cast<std::size_t>(th)];
+    std::fill(buf, buf + static_cast<std::size_t>(tw) * static_cast<std::size_t>(th), fill);
     return PixieData(1, static_cast<unsigned char>(tw),
                      static_cast<unsigned char>(th), buf);
 }
@@ -71,7 +71,7 @@ PixieData make_grid(int tw, int th, unsigned char fill)
 void paint(PixieData& g, int tx, int ty, unsigned char tile)
 {
     if (tx >= 0 && ty >= 0 && tx < g.w && ty < g.h)
-        g.data[tx + ty * g.w] = tile;
+        g.data[static_cast<std::size_t>(tx + ty * g.w)] = tile;
 }
 
 void paint_rect(PixieData& g, int tx0, int ty0, int tx1, int ty1,
@@ -110,7 +110,7 @@ bool paint_decor(GameWorld& w, int floor, int tx, int ty,
         return false;
     if (decor_id >= DECOR_MAX)
         return false;
-    const unsigned char base = g.data[tx + ty * g.w];
+    const unsigned char base = g.data[static_cast<std::size_t>(tx + ty * g.w)];
     if (base == PIX_AIR || base == PIX_ZSTAIR_UP || base == PIX_ZSTAIR_DOWN ||
         base == PIX_VOID1)
     {
@@ -128,7 +128,7 @@ bool paint_decor(GameWorld& w, int floor, int tx, int ty,
         dec = PixieData(1, static_cast<unsigned char>(g.w),
                         static_cast<unsigned char>(g.h), buf);
     }
-    dec.data[tx + ty * dec.w] = decor_id;
+    dec.data[static_cast<std::size_t>(tx + ty * dec.w)] = decor_id;
     return true;
 }
 
@@ -361,12 +361,12 @@ bool cell_standable(const GameWorld& world, int floor, int tx, int ty)
     const PixieData& g = world.grid_for_floor(floor);
     if (!g.valid() || tx < 0 || ty < 0 || tx >= g.w || ty >= g.h)
         return false;
-    if (!ground_cell_standable(g.data[tx + ty * g.w]))
+    if (!ground_cell_standable(g.data[static_cast<std::size_t>(tx + ty * g.w)]))
         return false;
     const PixieData& dec = world.decor_for_floor(floor);
     if (dec.valid() && dec.w == g.w && dec.h == g.h)
     {
-        const unsigned char d = dec.data[tx + ty * dec.w];
+        const unsigned char d = dec.data[static_cast<std::size_t>(tx + ty * dec.w)];
         if (d < DECOR_MAX &&
             kDecorRegistry[d].pass == DecorPassability::BlocksGround)
             return false;
@@ -387,7 +387,7 @@ bool cell_is_fall_landing(const GameWorld& w, int floor, int tx, int ty)
     {
         const PixieData& gg = w.grid_for_floor(g);
         if (!gg.valid() || tx >= gg.w || ty >= gg.h ||
-            gg.data[tx + ty * gg.w] != PIX_AIR)
+            gg.data[static_cast<std::size_t>(tx + ty * gg.w)] != PIX_AIR)
             return false; // solid above: no fall reaches this cell
         for (int dy = -1; dy <= 1; ++dy)
             for (int dx = -1; dx <= 1; ++dx)
@@ -423,7 +423,7 @@ void scatter_base_tiles(GameWorld& w, std::uint32_t seed, int floor, int tx0,
                 continue;
             if (x < 0 || y < 0 || x >= g.w || y >= g.h)
                 continue;
-            const unsigned char t = g.data[x + y * g.w];
+            const unsigned char t = g.data[static_cast<std::size_t>(x + y * g.w)];
             if (t == PIX_ZSTAIR_UP || t == PIX_ZSTAIR_DOWN || t == PIX_VOID1)
                 continue;
             if (cell_is_fall_landing(w, floor, x, y))
@@ -434,7 +434,7 @@ void scatter_base_tiles(GameWorld& w, std::uint32_t seed, int floor, int tx0,
                   variants[position_hash(seed, x, y, salt + kSaltVariant) % 4u]);
             PixieData& dec = w.decor_for_floor(floor);
             if (dec.valid() && x < dec.w && y < dec.h)
-                dec.data[x + y * dec.w] = DECOR_NONE;
+                dec.data[static_cast<std::size_t>(x + y * dec.w)] = DECOR_NONE;
         }
 }
 
@@ -529,7 +529,7 @@ void scatter_boulders(GameWorld& w, std::uint32_t seed, int floor, int tx0,
                 continue;
             if (x < 0 || y < 0 || x >= g.w || y >= g.h)
                 continue;
-            const unsigned char t = g.data[x + y * g.w];
+            const unsigned char t = g.data[static_cast<std::size_t>(x + y * g.w)];
             if (t == PIX_ZSTAIR_UP || t == PIX_ZSTAIR_DOWN || t == PIX_VOID1)
                 continue;
             if (blocks_weapons_or_flyers(t))
@@ -585,7 +585,7 @@ bool scatter_decor(GameWorld& w, std::uint32_t seed, int floor, int tx0,
                 continue;
             if (x < 0 || y < 0 || x >= g.w || y >= g.h)
                 continue;
-            const unsigned char t = g.data[x + y * g.w];
+            const unsigned char t = g.data[static_cast<std::size_t>(x + y * g.w)];
             bool allowed = false;
             for (const ScatterGround ground : grounds)
                 if (scatter_ground_matches(t, ground))
@@ -597,7 +597,7 @@ bool scatter_decor(GameWorld& w, std::uint32_t seed, int floor, int tx0,
                 continue;
             const PixieData& dec = w.decor_for_floor(floor);
             if (dec.valid() && x < dec.w && y < dec.h &&
-                dec.data[x + y * dec.w] != DECOR_NONE)
+                dec.data[static_cast<std::size_t>(x + y * dec.w)] != DECOR_NONE)
                 continue; // hand-placed decor keeps its cell
             if (cell_near_entity(w, floor, x, y, 0))
                 continue; // no one spawns standing in the set dressing
